backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
qrcode[pil]==7.4.2
websockets==12.0
python-dateutil==2.9.0
httpx==0.28.1
pytest-asyncio==0.24.0
pywin32>=306; sys_platform == "win32"

